Backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.0.0
